Reorganized files, added gamebuilder project
Created a more sane project and file structures. Put all the wrestler
attribute classes within the wrestler class, so that they are nested,
and so that they have access to private members of wrestler, which
solves some weird access issues I was going to have, while still leaving
things very encapsulated. These nested classes are all within their own
files so file bloat is less of an issue.

Added the gamebuilder project with some preliminary code, all pretty
messy. Midway through making this, I realized the actual Wrestler class
needed the refactoring above, to ease some of the access involved. So
that stopped. As it is, the project is running and nothing is broken, so
this is a nice rollback point if things head south.