Tons of functionality, error messages, etc
Added functionality for adding and removing things. Added new error
messages that need to be implemented so that people aren't changing
things in a way that breaks things.  Text needs editing, but enough work
has been done, and nothing is completely broken, so this commit has been
created as an unsafe rollback point.  GameMaker is still really bad
code, it needs tons of work.